refactor: remove API clients and fallback logic
- Remove wikidata/client.go (API client)
- Remove conceptnet/client.go (API client)
- Update serve.go to use local readers only, no fallback to API
- Add wikidata to required data sources list
- Move mapConceptNetRelation from client.go to reader.go

All data sources now require local SQLite indexes to be prepared
before the server can start. No external API dependencies.
